update mkdocs-material-dib for mkdocs 1.0
This merges a large number of upstream changes from mkdocs
to account for their 2->3 transition, which was caused by
mkdocs bumping from version 0.17 to version 1.0.

The main thing that changed was how navigation links were handled,
so the old mkdocs-material-dib theme would produce broken
navigation tables on any non-index page when the site was
generated with mkdocs 1.0.

This fixes the problem, but you must also upgrade
to mkdocs 1.0!